+++ src/client/App.jsx
@@ -20,6 +20,7 @@ import {
 import { ColorModeProvider } from './contexts/ColorModeContext.jsx';
 import { LoadingProvider } from './contexts/LoadingContext.jsx';
 import { GlobalStateProvider } from './contexts/GlobalStateContext.jsx';
+import { LocalizationProvider } from './contexts/LocalisationContext.jsx';
 
 export const App = () => {
     // check color mode preference
@@ -51,13 +52,15 @@ export const App = () => {
     // render jsx
     return(
         <GlobalStateProvider>
-            <LoadingProvider>
-                <ColorModeProvider value={colorMode}>
-                    <ThemeProvider theme={theme}>
-                        <div>agenda-display</div>
-                    </ThemeProvider>
-                </ColorModeProvider>
-            </LoadingProvider>
+            <LocalizationProvider>
+                <LoadingProvider>
+                    <ColorModeProvider value={colorMode}>
+                        <ThemeProvider theme={theme}>
+                            <div>agenda-display</div>
+                        </ThemeProvider>
+                    </ColorModeProvider>
+                </LoadingProvider>
+            </LocalizationProvider>
         </GlobalStateProvider>
     );
 }