+++ src/client/i18n/de/de.jsx
@@ -1,0 +1,13 @@
+/*
+ * This file is part of the agenda-display project.
+ *
+ * (c) Nils Bohrs
+ *
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
+'use strict';
+
+export const de = {
+    'Missing translation': 'Übersetzung fehlt',
+}